refactor proposal operation lookup routes
